// src/api/model/calendarTaskDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PriorityLevel } from "./priorityLevel";
import type { WorkSpaceTaskStatus } from "./workSpaceTaskStatus";

export interface CalendarTaskDto {
	id?: string;
	title?: string;
	/** @nullable */
	description?: string | null;
	start?: string;
	end?: string;
	allDay?: boolean;
	/** @nullable */
	color?: string | null;
	/** @nullable */
	backgroundColor?: string | null;
	/** @nullable */
	borderColor?: string | null;
	/** @nullable */
	textColor?: string | null;
	priority?: PriorityLevel;
	status?: WorkSpaceTaskStatus;
	/** @nullable */
	assignedToId?: string | null;
	/** @nullable */
	projectId?: string | null;
}